Updated Solution - GfG to GitHub

+++ Difficulty: Basic/Reverse digits/reverse-digits.cpp
@@ -1,11 +1,16 @@
-class Solution {
-  public:
-    int reverseDigits(int n) {
-        int r=0;
-        while(n){
-            r=r*10+n%10;
-            n/=10;
-        }
-        return r;
-    }
+class Solution
+{
+ public:
+ int reverseDigits(int n)
+ {
+     long long r = 0;
+     while(n)
+     {
+     r = r*10+n%10;
+     n/=10;
+     }
+     return r;    
+ }
+ 
+ 
 };
